♻️ refactor: consolidate resources under internal/resources/
Move plugins/tools/builtin → internal/resources (package resources)
and internal/embedded → internal/resources/binaries (package binaries).
Flat layout: skills/, souls/, subagents/, templates/ as data dirs under
internal/resources/, with binaries/ as its own package.
