+++ internal/agent/integration_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/vaayne/anna/internal/agent/runner"
-	"github.com/vaayne/anna/internal/embedded"
+	"github.com/vaayne/anna/internal/resources/binaries"
 	"github.com/vaayne/anna/pkg/providers"
 	anthropicprovider "github.com/vaayne/anna/plugins/providers/anthropic"
 )
@@ -35,7 +35,7 @@ func TestIntegrationPoolWithGoRunner(t *testing.T) {
 	if err := os.MkdirAll(binDir, 0o755); err != nil {
 		t.Fatalf("MkdirAll: %v", err)
 	}
-	_ = embedded.EnsureTools(annaHome)
+	_ = binaries.EnsureTools(annaHome)
 	boxshPath := filepath.Join(binDir, "boxsh")
 	_ = os.Remove(boxshPath)
 	if err := os.WriteFile(boxshPath, []byte("#!/bin/sh\necho boxsh 2.0.1\n"), 0o755); err != nil {
