+++ internal/resources/binaries/binaries/windows-amd64/PLACEHOLDER
