+++ internal/resources/binaries/embed_linux_amd64.go
@@ -1,6 +1,6 @@
 //go:build linux && amd64
 
-package embedded
+package binaries
 
 import "embed"
