+++ internal/resources/binaries/embed_linux_arm64.go
@@ -1,6 +1,6 @@
 //go:build linux && arm64
 
-package embedded
+package binaries
 
 import "embed"
