+++ internal/resources/binaries/embed_windows_arm64.go
@@ -1,6 +1,6 @@
 //go:build windows && arm64
 
-package embedded
+package binaries
 
 import "embed"
