+++ internal/resources/binaries/embedded.go
@@ -1,4 +1,4 @@
-package embedded
+package binaries
 
 import (
 	"bytes"
