+++ internal/resources/binaries/embedded_test.go
@@ -1,4 +1,4 @@
-package embedded
+package binaries
 
 import (
 	"os"
