+++ internal/resources/embed.go
@@ -1,7 +1,7 @@
 // Package builtin bundles embedded resources (skills, souls, sub-agents, templates)
 // that Anna ships with its binary. Runtime code accesses them through Registry,
 // not by walking the filesystem directly.
-package builtin
+package resources
 
 import (
 	"embed"
