+++ internal/resources/registry.go
@@ -1,4 +1,4 @@
-package builtin
+package resources
 
 import (
 	"errors"
