+++ internal/resources/registry_test.go
@@ -1,4 +1,4 @@
-package builtin
+package resources
 
 import (
 	"testing"
