+++ internal/resources/skills/anna/references/builtin-resources.md
