+++ internal/resources/skills/anna/references/configuration.md
