+++ internal/resources/skills/docs-writing/SKILL.md
