+++ internal/resources/skills/implementation/SKILL.md
