+++ internal/resources/skills/research/SKILL.md
