+++ internal/resources/souls/default.md
