+++ internal/resources/souls/direct.md
