+++ internal/resources/souls/researcher.md
