+++ internal/resources/souls/teacher.md
