+++ internal/resources/subagents/researcher.md
