+++ internal/resources/subagents/writer.md
