+++ internal/resources/templates/default.md
