+++ internal/resources/templates/researcher.md
