+++ internal/resources/templates/writer.md
